Split model-state I/O out of tts_model; collapse int | float hints
