+++ pocket_tts/__init__.py
@@ -8,11 +8,14 @@ if os.environ.get("POCKET_TTS_NO_BEARTYPE") != "1":
     from beartype import BeartypeConf
     from beartype.claw import beartype_this_package
 
-    beartype_this_package(conf=BeartypeConf(is_color=False))
+    # is_pep484_tower: accept int where float is annotated, as PEP 484 says.
+    # Without it an `int | float` hint is not redundant and ruff's PYI041
+    # advice to collapse it would break callers passing ints.
+    beartype_this_package(conf=BeartypeConf(is_color=False, is_pep484_tower=True))
 
+from pocket_tts.models.model_state import export_model_state
 from pocket_tts.models.tts_model import (  # noqa: E402
     TTSModel,
-    export_model_state,
 )
 
 # Public methods:
